requirements.txt
numpy==1.19.3
Flask==2.0.1
requests==2.22.0
imageai==2.1.6
Pillow==8.3.2
